no.9/requirements.txt
fastapi
uvicorn
python-dotenv
openai
tweepy
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic
apscheduler
